+++ linx/utils/types/shared-types/browsingImage.ts
@@ -1,0 +1,17 @@
+
+export interface BrowsingImage {
+    Attributes: unknown[];
+    HasMediaPath: boolean;
+    Height: number;
+    Index: number;
+    MaxHeight: number;
+    MaxWidth: number;
+    MediaPath: string;
+    MediaSizeType: string;
+    MediaType: string;
+    ProductPath: string;
+    Title: null;
+    Url: null;
+    VariationPath: string;
+    Width: number;
+  }
