+++ linx/utils/types/shared-types/productDefinition.ts
@@ -1,9 +1,8 @@
 export interface ProductDefinition {
-    IsJoinable: boolean;
-    JoinDisplayText: null;
-    JoinProductBehavior: number;
-    JoinPropertyMetadataID: null;
-    JoinStockKeepUnitPatternLength: null;
-    ProductDefinitionID: number;
-  }
-  
+  IsJoinable: boolean;
+  JoinDisplayText: null;
+  JoinProductBehavior: number;
+  JoinPropertyMetadataID: null;
+  JoinStockKeepUnitPatternLength: null;
+  ProductDefinitionID: number;
+}
