+++ linx/utils/types/shared-types/productSelection.ts
@@ -1,8 +1,8 @@
 export type Path = "/";
 
 export interface ProductSelection {
-    PropertyLevel: number;
-    PropertyPath: Path;
-    Quantity: number;
-    SkuID: number;
-  }
+  PropertyLevel: number;
+  PropertyPath: Path;
+  Quantity: number;
+  SkuID: number;
+}
